requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
surya-ocr>=0.4.0
opencv-python>=4.8.0
Pillow>=10.0.0
jinja2>=3.1.0
aiofiles>=23.0.0
numpy>=1.24.0
requests>=2.31.0
transformers>=4.36.0
torch>=2.0.0
scipy>=1.10.0
